src/test/java/io/github/softwarecats/roulette/WheelTest.java
/*
 * MIT License
 *
 * Copyright © 2021 Bowen Wu. All rights reserved.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package io.github.softwarecats.roulette;

import org.junit.Before;
import org.junit.Test;

import java.util.Random;

import static org.junit.Assert.assertEquals;

public class WheelTest {

    Wheel wheel;

    @Before
    public void setUp() {
        Random rng = new Random(1);
        wheel = new Wheel(rng);
    }

    @Test
    public void addOutcome() {
        Outcome outcome = new Outcome("1", 1);
        Bin bin = (Bin) wheel.getBin(1).clone();

        wheel.addOutcome(1, outcome);
        bin.add(outcome);

        assertEquals(wheel.getBin(1), bin);
    }

    @Test
    public void next() {
        Random rng = new Random(1);

        assertEquals(wheel.next(), wheel.getBin(rng.nextInt(38)));
    }

    @Test
    public void getOutcomes() {
        BinBuilder builder = new BinBuilder();
        builder.buildBins(wheel);

        assertEquals(wheel.getOutcomes("Line").size(), 11);
    }
}